+++ gui/views/route/route_view.py
@@ -1,7 +1,7 @@
 import ttkbootstrap as tb
 from ttkbootstrap.constants import *
 from netkit.utils.ui_helper import ui_helper
-from netkit.services.route import RouteService
+from netkit.services.route.route import RouteService
 import tkinter.messagebox as mbox
 import threading
 from tkinter import ttk